+++ assets/components/modxsite/templates/shopmodx/v4.0/app/modules/Site/components/App/index.js
@@ -12,6 +12,10 @@ import Renderer from './Renderer';
 
 import Basket from '../Basket';
 
+
+import defaultQuery from '../ORM/query';
+import rootResolver from '../ORM/resolver';
+
 const {
 	mapDispatchToProps,
 	mapStateToProps,