+++ assets/components/modxsite/templates/shopmodx/v4.0/internals/webpack/webpack.base.babel.js
@@ -39,6 +39,7 @@ module.exports = (options) => ({
 							|| /\/react-draft-wysiwyg\//.test(file)
 							|| /\/shopmodx-react\//.test(file)
 							|| /\/modx-react\//.test(file)
+							|| /\/react-cms-graphql-utils\//.test(file)
 						){
 							return;
 						}