+++ core/components/modxsite/processors/web/orders/own/object.class.php
@@ -1,0 +1,9 @@
+<?php
+
+require_once MODX_CORE_PATH . 'components/shopmodx/processors/shopmodx/orders/own/object.class.php';
+
+class modWebOrdersOwnObjectProcessor extends modShopmodxOrdersOwnObjectProcessor{
+
+}
+
+return 'modWebOrdersOwnObjectProcessor';