+++ core/components/modxsite/processors/web/users/create.class.php
@@ -8,14 +8,13 @@
 
 require_once MODX_PROCESSORS_PATH . 'security/user/create.class.php';
 
-require_once dirname(__FILE__) . '/_validator.class.php';
-
-class modWebSocietyUsersCreateProcessor extends modUserCreateProcessor{
+class modWebUsersCreateProcessor extends modUserCreateProcessor{
     
     public $permission = '';
     
     public function initialize(){
 
+        
         $this->setDefaultProperties(array(
             "passwordnotifymethod"  => "this",
             "passwordgenmethod"     => "g",
@@ -27,12 +26,12 @@ class modWebSocietyUsersCreateProcessor extends modUserCreateProcessor{
             "use_captcha"           => false,
             "groups"                => array(
                 array(
-                    "usergroup" => 12,
+                    "usergroup" => 10,
                     "role"      => 1,
                 ),
             ),
-            'email'                 => mb_convert_case($this->getProperty('email'), MB_CASE_LOWER, 'utf-8'),
-            "mail_template"         => "messages/users/registration.tpl",
+            'email'                 => mb_convert_case(trim($this->getProperty('email')), MB_CASE_LOWER, 'utf-8'),
+            "mail_template"         => "message/users/registration.tpl",
         ));
         
         
@@ -116,19 +115,12 @@ class modWebSocietyUsersCreateProcessor extends modUserCreateProcessor{
          
         $user =& $this->object;
         
-        if(!isset($user->delegate)){
-            $this->addFieldError("delegate", "Укажите являетесь ли вы представителем компании");
-        }
-
-        $user->SocietyProfile = $this->modx->newObject('SocietyUserProfile', array(
-            "createdon" => time(),
-        ));
-        
         if($this->getProperty('use_captcha') && !$user->SocialProfiles){
             $this->checkCaptcha(); 
         } 
         
         $ok = parent::beforeSave();
+
         if($ok !== true){
             return $ok;
         } 
@@ -137,42 +129,17 @@ class modWebSocietyUsersCreateProcessor extends modUserCreateProcessor{
         $profile->fromArray(array(
             "fullname"  => $this->getProperty('fullname'),
         ));
-         
-        // else
-        
-        $validator = new modWebSocietyUsersValidator($this);
-        
-        $ok = $validator->validate();
-        
-        if($ok !== true){
-            return $ok;
-        }
-        
-        # print '<pre>';
         
         // Добавляем сразу все типы уведомлений
-        $notices = array();
-        foreach($this->modx->getCollection('SocietyNoticeType') as $noticeType){
-            $notice = $this->modx->newObject('SocietyNoticeUser');
-            $notice->NoticeType = $noticeType;
+        // $notices = array();
+        // foreach($this->modx->getCollection('SocietyNoticeType') as $noticeType){
+        //     $notice = $this->modx->newObject('SocietyNoticeUser');
+        //     $notice->NoticeType = $noticeType;
             
-            # print_r($notice->toArray());
-            $notices[] = $notice;
-        }
-        $user->Notices = $notices;
-        
-        /*
-        
-        return "Ведутся работы";
-        
-        print_r($this->properties);
-        
-        print $activate_link;
-            
-        */
-
-        // print_r($user->toArray());
-        // return "Debug";
+        //     # print_r($notice->toArray());
+        //     $notices[] = $notice;
+        // }
+        // $user->Notices = $notices;
         
         return true;
     }
@@ -202,15 +169,8 @@ class modWebSocietyUsersCreateProcessor extends modUserCreateProcessor{
             }
             
             // Генерация ссылки для активации
-            // $activate_link = $this->modx->makeUrl(
-            //     998, '', array(
-            //         "u"     => $user->id,
-            //         "k"     => md5($user->id . $user->Profile->email),
-            //     ),
-            //     'full'
-            // );
             
-            $activation_key = md5($user->id . $user->Profile->email);
+            $activation_key = md5($this->modx->site_id . $user->id . $user->Profile->email);
 
             $activate_link = $this->modx->getOption("site_url") . "profile/{$user->username}/activation/?k={$activation_key}";
 
@@ -222,47 +182,6 @@ class modWebSocietyUsersCreateProcessor extends modUserCreateProcessor{
             $this->modx->smarty->assign('properties', $this->getProperties());
             $this->modx->smarty->assign('user', $user);
 
-
-
-//             if($this->getProperty("isOffer")){
-                
-//                 $date = date("m-d-Y", time() + (3600 * 24 * 7));
-                
-//                 $offer = <<<HTML
-//                     <p>
-                        
-//                         Мы запустили обновленную версию нашего проекта: <a href="http://gorodskie-bani.ru/">http://gorodskie-bani.ru/</a>
-
-//                     </p>
-
-//                     <p>
-//                         Теперь портал будет развиваться в формате глобального картографического сервиса с направленностью на обширную аудиторию. 
-//                     </p>
-
-//                     <p>
-//                         В течение месяца будет опубликован функционал для самостоятельного добавления бань и саун на карте (что должно обеспечить значительную генерацию контента и увеличение посещаемости, вовлеченности и т.п.). Но будут вводиться и платные функции (индивидуальные иконки, расширенные карточки организаций, онлайн-формы, аналитика и т.п.). По аналитике у нас будут вообще очень интересные решения вплоть до возможности просмотреть в каких районах карты какое количество просмотров пользователей и т.п., а так же возможность видеть пользователей онлайн и возможность инициировать диалог с пользователем. Сейчас мы планируем вводить оплату за каждую услугу в отдельности, чтобы каждый сам решал какие бюджеты выделять на этот рекламный канал. По нашим оценкам средний чек нашего партнера будет составлять 3 - 10 тысяч рублей в месяц уже через 3 месяца (когда мы введем бОльшую часть запланированного функционала). Но на развитие проекта финансы нужны уже сейчас, поэтому у нас есть для Вас спецпредложение: премиальное размещение на год всего за 15 000 рублей разово. При оплате до {$date} мы разместим Вашу карточку, присвоим Вашей компании индивидуальную иконку на карте, а так же гарантируем предоставление всех вводимых на портале услуг в течение этого года бесплатно. Плюс ко всему этому гарантируется наше особое отношение к Вашей компании и максимальная лояльность во всем, включая пожелания к развитию проекта (кроме политики пользовательских отзывов и т.п., с этом у нас очень серьезно все.). Уверяю, это очень выгодное предложение. Вот динамика посещения портала: <a href="http://joxi.ru/Drlz7Mqc4JdNL2">http://joxi.ru/Drlz7Mqc4JdNL2</a> На этот новый год, уверен, будет более 2000 пользователей, а к следующему НГ не менее 10 000 (так как сейчас будет объектов добавляться много, соответственно больше материала для поискового трафика будет). У нас весь трафик белый, не покупной. Могу предоставить доступ в Яндекс.Метрику для просмотра статистики.
-//                     </p>
-                        
-
-//                     <p>
-                    
-//                         Возможна оплата по договору.
-                        
-//                     </p>
-// HTML;
-
-//                 $this->modx->smarty->assign('text', $offer);
-
-
-//                 $user->fromArray(array(
-//                     "offer" => $offer,
-//                     "offer_date"    => time(),
-//                     "createdon"     => 0,
-//                 ));
-
-//                 $user->save();
-
-//             }
 
             $message = $this->modx->smarty->fetch($mail_template);
             $subject = "Регистрация на сайте {$site_name}";
@@ -311,5 +230,5 @@ class modWebSocietyUsersCreateProcessor extends modUserCreateProcessor{
 }
 
 
-return 'modWebSocietyUsersCreateProcessor';
+return 'modWebUsersCreateProcessor';