+++ assets/components/modxsite/templates/shopmodx/v4.0/CHANGELOG.md
@@ -1,4 +1,23 @@
 
+4.3.0
+============================================
+- Extends react-crm defaultQuery by shopmodx-react defaultQuery
+- Added usersList query
+- Added ordersList query
+- Added usersOrdersList query
+- Added userById query
+- Added userByUsername query
+- Added users registration
+- Added user page
+- Added mutation updateUser
+- Added uploadImage
+- Added Orders Page
+
+
+4.2.1
+============================================
+- Added defaultQuery extension via react-cms-graphql-utils
+
 4.2.0
 ============================================
 - Upgrade packages: